feat: Convert knex, sendgrid and stripe connections to plugin structure.
