+++ packages/plugins/connections/connection-knex/src/index.js
@@ -1,0 +1,7 @@
+import Knex from './connections/Knex/Knex.js';
+
+export const connections = {
+  Knex,
+};
+
+export default { connections };
