+++ packages/plugins/connections/connection-sendgrid/src/connections/SendGridMail/SendGridMailSchema.json
