+++ packages/plugins/connections/connection-sendgrid/src/connections/SendGridMail/SendGridMailSend/SendGridMailSendSchema.json
