+++ packages/plugins/connections/connection-sendgrid/src/index.js
@@ -1,0 +1,7 @@
+import SendGridMail from './connections/SendGridMail/SendGridMail.js';
+
+export const connections = {
+  SendGridMail,
+};
+
+export default { connections };
