+++ packages/plugins/connections/connection-stripe/jest.config.cjs
