+++ packages/plugins/connections/connection-stripe/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lowdefy/connection-mongodb",
+  "name": "@lowdefy/connection-stripe",
   "version": "3.22.0",
   "licence": "Apache-2.0",
   "description": "",
@@ -38,16 +38,15 @@
     "clean": "rm -rf dist",
     "prepare": "yarn build",
     "swc": "swc src --out-dir dist --config-file ../../../../.swcrc --delete-dir-on-start --copy-files",
-    "test": "jest --coverage"
+    "test": "jest --coverage",
+    "s": "swc"
   },
   "dependencies": {
     "@lowdefy/helpers": "3.22.0",
-    "mongodb": "3.7.3",
-    "saslprep": "1.0.3"
+    "stripe": "8.186.0"
   },
   "devDependencies": {
     "@lowdefy/ajv": "3.22.0",
-    "@shelf/jest-mongodb": "2.1.0",
     "@swc/cli": "0.1.51",
     "@swc/core": "1.2.107",
     "@swc/jest": "0.2.5",
