+++ packages/plugins/connections/connection-stripe/src/connections/Stripe/StripeRequest/StripeRequestSchema.json
