+++ packages/plugins/connections/connection-stripe/src/connections/Stripe/StripeSchema.json
