+++ packages/plugins/connections/connection-stripe/src/index.js
@@ -1,0 +1,7 @@
+import Stripe from './connections/Stripe/Stripe.js';
+
+export const connections = {
+  Stripe,
+};
+
+export default { connections };
