--- packages/plugins/connections/knex/src/Knex/Knex.test.js
@@ -1,60 +1,0 @@
-/*
-  Copyright 2020-2021 Lowdefy, Inc
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-import { validate } from '@lowdefy/ajv';
-import Knex from './Knex.js';
-
-const { schema } = Knex;
-
-test('All requests are present', () => {
-  expect(Knex.requests.KnexRaw).toBeDefined();
-  expect(Knex.requests.KnexBuilder).toBeDefined();
-});
-
-test('valid connection schema, string connection', () => {
-  const connection = {
-    client: 'pg',
-    connection: 'postgresql://user:password@database.server.com:5432/db',
-  };
-  expect(validate({ schema, data: connection })).toEqual({ valid: true });
-});
-
-test('valid connection schema, object connection', () => {
-  const connection = {
-    client: 'pg',
-    connection: {
-      host: 'database.server.com',
-      user: 'user',
-      password: 'password',
-      database: 'db',
-    },
-  };
-  expect(validate({ schema, data: connection })).toEqual({ valid: true });
-});
-
-test('client missing', () => {
-  const connection = { connection: 'postgresql://user:password@database.server.com:5432/db' };
-  expect(() => validate({ schema, data: connection })).toThrow(
-    'Knex connection should have required property "client".'
-  );
-});
-
-test('client missing', () => {
-  const connection = { client: 'pg' };
-  expect(() => validate({ schema, data: connection })).toThrow(
-    'Knex connection should have required property "connection".'
-  );
-});
