--- packages/plugins/connections/knex/src/index.js
@@ -1,7 +1,0 @@
-import Knex from './Knex/Knex';
-
-export const connections = {
-  Knex,
-};
-
-export default { connections };
