--- packages/plugins/connections/sendgrid/src/index.js
@@ -1,7 +1,0 @@
-import SendGridMail from './SendGridMail/SendGridMail.js';
-
-export const connections = {
-  SendGridMail,
-};
-
-export default { connections };
