--- packages/plugins/connections/stripe/src/index.js
@@ -1,7 +1,0 @@
-import Stripe from './Stripe/Stripe.jss';
-
-export const connections = {
-  Stripe,
-};
-
-export default { connections };
